Move CacheSize component to own component file
